Move parser spi (#7975)
