+++ shardingsphere-sql-parser/shardingsphere-sql-parser-dialect/shardingsphere-sql-parser-oracle/src/main/java/org/apache/shardingsphere/sql/parser/oracle/parser/OracleParserFacade.java
@@ -15,31 +15,29 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sqlserver;
+package org.apache.shardingsphere.sql.parser.oracle.parser;
 
-import org.apache.shardingsphere.sql.parser.api.lexer.SQLLexer;
+import org.apache.shardingsphere.sql.parser.api.parser.SQLLexer;
 import org.apache.shardingsphere.sql.parser.api.parser.SQLParser;
 import org.apache.shardingsphere.sql.parser.spi.SQLParserFacade;
-import org.apache.shardingsphere.sql.parser.sqlserver.lexer.SQLServerLexer;
-import org.apache.shardingsphere.sql.parser.sqlserver.parser.SQLServerParser;
 
 /**
- * SQL parser facade for SQLServer.
+ * SQL parser facade for Oracle.
  */
-public final class SQLServerParserFacade implements SQLParserFacade {
+public final class OracleParserFacade implements SQLParserFacade {
     
     @Override
     public String getDatabaseType() {
-        return "SQLServer";
+        return "Oracle";
     }
     
     @Override
     public Class<? extends SQLLexer> getLexerClass() {
-        return SQLServerLexer.class;
+        return OracleLexer.class;
     }
     
     @Override
     public Class<? extends SQLParser> getParserClass() {
-        return SQLServerParser.class;
+        return OracleParser.class;
     }
 }
